+++ rewatch/src/build.rs
@@ -8,6 +8,8 @@ pub mod packages;
 pub mod parse;
 pub mod read_compile_state;
 
+use self::compile::compiler_args;
+use self::parse::parser_args;
 use crate::build::compile::{mark_modules_with_deleted_deps_dirty, mark_modules_with_expired_deps_dirty};
 use crate::helpers::emojis::*;
 use crate::helpers::{self, get_workspace_root};
@@ -25,9 +27,6 @@ use std::io::{Write, stdout};
 use std::path::{Path, PathBuf};
 use std::process::Stdio;
 use std::time::{Duration, Instant};
-
-use self::compile::compiler_args;
-use self::parse::parser_args;
 
 fn is_dirty(module: &Module) -> bool {
     match module.source_type {
@@ -56,7 +55,7 @@ pub struct CompilerArgs {
     pub parser_args: Vec<String>,
 }
 
-pub fn get_compiler_args(path: &Path, build_dev_deps: bool) -> Result<String> {
+pub fn get_compiler_args(path: &Path) -> Result<String> {
     let filename = &helpers::get_abs_path(path);
     let package_root =
         helpers::get_abs_path(&helpers::get_nearest_config(&path).expect("Couldn't find package root"));
@@ -64,6 +63,10 @@ pub fn get_compiler_args(path: &Path, build_dev_deps: bool) -> Result<String> {
     let root_rescript_config =
         packages::read_config(&workspace_root.to_owned().unwrap_or(package_root.to_owned()))?;
     let rescript_config = packages::read_config(&package_root)?;
+    let is_type_dev = match filename.strip_prefix(&package_root) {
+        Err(_) => false,
+        Ok(relative_path) => root_rescript_config.find_is_type_dev_for_path(relative_path),
+    };
 
     // make PathBuf from package root and get the relative path for filename
     let relative_filename = filename.strip_prefix(PathBuf::from(&package_root)).unwrap();
@@ -97,7 +100,7 @@ pub fn get_compiler_args(path: &Path, build_dev_deps: bool) -> Result<String> {
         &package_root,
         &workspace_root,
         &None,
-        build_dev_deps,
+        is_type_dev,
         true,
     );
 
@@ -281,7 +284,6 @@ pub fn incremental_build(
     show_progress: bool,
     only_incremental: bool,
     create_sourcedirs: bool,
-    build_dev_deps: bool,
     snapshot_output: bool,
 ) -> Result<(), IncrementalBuildError> {
     logs::initialize(&build_state.packages);
@@ -393,7 +395,6 @@ pub fn incremental_build(
         show_progress,
         || pb.inc(1),
         |size| pb.set_length(size),
-        build_dev_deps,
     )
     .map_err(|e| IncrementalBuildError {
         kind: IncrementalBuildErrorKind::CompileError(Some(e.to_string())),
@@ -500,7 +501,6 @@ pub fn build(
         show_progress,
         false,
         create_sourcedirs,
-        build_dev_deps,
         snapshot_output,
     ) {
         Ok(_) => {